Rename modules - glacierBer -> glacier_bed - glacierCollection -> glacier_collection - massBalance -> mass_balance
Also updates relevant import statements.
