+++ oggm_edu/__init__.py
@@ -2,9 +2,9 @@ __version__ = "0.1.0b"
 
 from oggm_edu.funcs import plot_glacier_graphics, initalize_oggm
 from oggm_edu.glacier import Glacier, SurgingGlacier
-from oggm_edu.glacierBed import GlacierBed
-from oggm_edu.glacierCollection import GlacierCollection
-from oggm_edu.massBalance import MassBalance
+from oggm_edu.glacier_bed import GlacierBed
+from oggm_edu.glacier_collection import GlacierCollection
+from oggm_edu.mass_balance import MassBalance
 
 # Initialize on import
 initalize_oggm()
